Added header parsing support, redid error system.
It can now parse the header and the status line of the HTTP package,
headers are parsed into a hashmap from stdlib. Also redid error handling
by switching to an ENUM to allow diffrent errors in this commit + the
previous commit. Also added error handling for all parts of the header
parsing setup.

To-do:
- Switch everything over to new error system
- Implement proper display for error system to allow for actual
  debuggining.. I know crazy idear.
- Implement response
- GAME